Fixed ci.yaml jobs
