--- redis-e2e-test-sink/README.md
@@ -1,4 +1,0 @@
-# Redis E2E Test Sink
-A User Defined Sink that set the payload of each incoming datum element as a key in a redis instance with dummy value 1.
-This sink is used by E2E testing.
-
