--- redis-e2e-test-sink/go.mod
@@ -1,23 +1,0 @@
-module github.com/numaproj/numaflow-sinks/argoworkflow
-
-go 1.18
-
-require (
-	github.com/go-redis/redis/v8 v8.11.4
-	github.com/numaproj/numaflow-go v0.5.1
-)
-
-require (
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/fsnotify/fsnotify v1.5.1 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/stretchr/testify v1.8.1 // indirect
-	golang.org/x/net v0.9.0 // indirect
-	golang.org/x/sys v0.7.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
-	google.golang.org/genproto v0.0.0-20230526161137-0005af68ea54 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20230525234030-28d5490b6b19 // indirect
-	google.golang.org/grpc v1.57.0 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
